Adds the Stanford parser, implementation to TweetCleaner, adjustments to Streamer
Atomicity is employed in Streamer to make the filterUntil method reliable (it actually works!). Performance impact is negligible.

You only need to run the program once now, since it retrieves all emotions and inserts them into the database.

Adds Stanford's PTB tokenizer to tokenize tweets. Still doesn't filter English, but maybe we'll worry about that later. (The upside: on the front end, we'll be able to "classify" non-English tweets that will come in… but still :|)

Signed-off-by: Kevin Panuelos <kevin.panuelos@gmail.com>